Centralize value→JSON serialization on Value::to_json()
Five copies of nearly-identical `fn value_to_json(v: &Value) -> serde_json::Value`
existed across crates (lex-cli, lex-runtime, lex-api, spec-checker,
conformance) plus the original in lex-trace's recorder. Each had
to grow Map/Set arms when those Value variants landed last round,
which is the kind of multiplication that turns into bit-rot under
new variants.

Add `Value::to_json()` to lex-bytecode and replace the five
duplicates with one-line wrappers. Trace's recorder keeps its
specialized form (it uses `$map`/`$set`/`$variant`/`$f64_array`
markers so traces are fully reversible — `lex replay` reads them
back into Values). The five shared sites all want the
"native JSON" form (Str-keyed map → object, otherwise array of
pairs); that's what `Value::to_json` emits.

Net diff: -204 / +88. Same workspace test count (282), same
clippy-clean.
